WIP: écran race

+++ app/src/main/java/fr/btytgat/odysseedesvagabonds/database/entities/Race.kt
@@ -4,10 +4,10 @@ package fr.btytgat.odysseedesvagabonds.database.entities
 data class Race(
     var uuid: String,
     var name: String = "",
-    var healthDice: Int,
-    var manaDice: Int,
+    var healthDice: Long,
+    var manaDice: Long,
     var description: String = "",
     var history: String = "",
     var uuidVoie: String,
-    var statsChange: HashMap<String, Int>
+    var statsChange: HashMap<String, Long>
 )
